requirements.txt
beautifulsoup4==4.12.2
chromadb==0.4.18
lxml==4.9.3
PyYAML==6.0.1
Requests==2.31.0
streamlit>=1.28.0
streamlit-lottie>=0.0.5